requirements.txt
fastapi
openai
cryptography
pydantic
requests
pillow
uvicorn